backend/requirements.txt
django==3.0.5
dnspython==1.16.0
mongoengine==0.20.0
pymongo==3.10.1
djangorestframework==3.12.4